requirements.txt
torch 
torchaudio 
transformers 
librosa 
dtw-python 
phonemizer 
numpy
fastapi
python-multipart
coqui-tts